# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bancaire - Gestion des Comptes",
        "description": "Documentation complète de l'API de Gestion des Comptes avec exemples de test. Utilisez les credentials suivants pour tester : Admin (admin@gestioncomptes.com / password) ou Client (client@gestioncomptes.com / password)",
        "contact": {
            "email": "contact@gestioncomptes.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://gestioncompteslaravel.onrender.com",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000/api/v1/ndiaye.mapathe",
            "description": "Local Development"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion utilisateur",
                "description": "Authentifier un utilisateur et retourner un token d'accès JWT",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email de l'utilisateur",
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@gestioncomptes.com"
                                    },
                                    "password": {
                                        "description": "Mot de passe de l'utilisateur",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Admin User"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "admin@gestioncomptes.com"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "admin"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9..."
                                                },
                                                "refresh_token": {
                                                    "type": "string",
                                                    "example": "token_id_here"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "expires_in": {
                                                    "type": "integer",
                                                    "example": 604800
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "INVALID_CREDENTIALS"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Email ou mot de passe incorrect"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/comptes/archives": {
            "get": {
                "tags": [
                    "Comptes Archivés"
                ],
                "summary": "Lister tous les comptes archivés",
                "description": "Récupérer la liste de tous les comptes archivés. La consultation de compte Epargne archiver se fait à partir du cloud",
                "operationId": "listComptesArchives",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type de compte",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "epargne",
                                "cheque",
                                "courant"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "actif",
                                "bloque",
                                "ferme"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Rechercher par numéro de compte ou nom du titulaire",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "dateCreation",
                                "solde",
                                "titulaire"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre de tri",
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes archivés récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des comptes archivés récupérée avec succès"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/clients/{client}/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Récupérer les comptes d'un client",
                "description": "Récupérer tous les comptes associés à un client spécifique",
                "operationId": "getClientComptes",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "ID du client",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type de compte. Peut être une valeur unique ou un tableau de valeurs (epargne, cheque, courant)",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "epargne",
                                    "cheque",
                                    "courant"
                                ]
                            },
                            "example": [
                                "epargne",
                                "cheque"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut du compte",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "actif",
                                "bloque",
                                "ferme"
                            ],
                            "example": "actif"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Rechercher par numéro de compte ou nom du titulaire",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "COMP-123456"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "dateCreation",
                                "solde",
                                "titulaire"
                            ],
                            "example": "dateCreation"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "desc"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comptes du client récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comptes du client récupérés avec succès"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "currentPage": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "totalPages": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "totalItems": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "itemsPerPage": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "hasNext": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "hasPrevious": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "links": {
                                            "properties": {
                                                "self": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/v1/ndiaye.mapathe/clients/550e8400-e29b-41d4-a716-446655440000/comptes?page=1"
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/v1/ndiaye.mapathe/clients/550e8400-e29b-41d4-a716-446655440000/comptes?page=2"
                                                },
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/v1/ndiaye.mapathe/clients/550e8400-e29b-41d4-a716-446655440000/comptes?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/v1/ndiaye.mapathe/clients/550e8400-e29b-41d4-a716-446655440000/comptes?page=5"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/clients/{client}/comptes/{compte}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Récupérer un compte spécifique d'un client",
                "description": "Récupérer les détails d'un compte spécifique appartenant à un client",
                "operationId": "getClientCompte",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "ID du client",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "compte",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte récupéré avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client ou compte non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Client": {
                "title": "Client",
                "description": "Modèle de client",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "titulaire": {
                        "type": "string",
                        "example": "Mamadou Diop"
                    },
                    "nci": {
                        "type": "string",
                        "pattern": "^[12]\\\\d{12}$",
                        "example": "1234567890123"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "mamadou.diop@email.com"
                    },
                    "telephone": {
                        "type": "string",
                        "pattern": "^\\\\+221[0-9]{9}$",
                        "example": "+221776543210"
                    },
                    "adresse": {
                        "type": "string",
                        "example": "123 Rue Kermel, Plateau, Dakar"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Compte": {
                "title": "Compte",
                "description": "Modèle de compte bancaire",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    },
                    "numero": {
                        "type": "string",
                        "example": "COMP-123456"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "epargne",
                            "cheque",
                            "courant"
                        ],
                        "example": "cheque"
                    },
                    "soldeInitial": {
                        "type": "number",
                        "format": "double",
                        "example": 50000
                    },
                    "solde": {
                        "type": "number",
                        "format": "double",
                        "example": 45000
                    },
                    "devise": {
                        "type": "string",
                        "enum": [
                            "XOF",
                            "FCFA",
                            "USD",
                            "EUR"
                        ],
                        "example": "XOF"
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "actif",
                            "bloque",
                            "ferme"
                        ],
                        "example": "actif"
                    },
                    "client_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "archived": {
                        "type": "boolean",
                        "example": false
                    },
                    "motifBlocage": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "client": {
                        "$ref": "#/components/schemas/Client"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}